# soy/exercise-http/doc/API/openApiDocumentation.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Plage API",
    "description": "Plage API \n \n **/!\\ Complex types such as *Dates* or *bytea* are labeled as *String* in this documentation. The given string's format describes what complex data type it stands for. You can retrieve these details in Models' description section down below.** ",
    "termsOfService": "None",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org"
    }
  },
  "basePath": "/",
  "paths": {
...
    "/api/exercise": {
      "post": {
        "tags": ["Exercise"],
        "summary": "Create an exercise",
        "description": "**Authorized roles [Teacher, Admin]**",
        "operationId": "devPostExercise",
        "consumes": [],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Exercise",
            "in": "body",
            "description": "An exercise",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["state", "name", "locale"],
              "properties": {
                "template_statement": {
                  "type": "string"
                },
                "template_archive": {
                  "type": "string",
                  "format": "bytea"
                },
                "state": {
                  "type": "string"
                },
                "author": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "statement_creation_script": {
                  "type": "string",
                  "format": "bytea"
                },
                "marking_script": {
                  "type": "string",
                  "format": "bytea"
                },
                "locale": {
                  "type": "string",
                  "format": "loc"
                },
                "ref_id": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Exercise id",
            "schema": {
              "type": "object",
              "properties": {
                "ex_id": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Error: Invalid request"
          },
          "401": {
            "description": "Error: Autorization required"
          },
          "500": {
            "description": "Error: Internal Server Error"
          }
        }
      }
    },
...
    "/api/exercise-production": {
      "post": {
        "tags": ["Exercise Production"],
        "summary": "Creates a new exercise production",
        "description": "*Different route already implemented* - Creates a new exercise production\n **Authorized roles : [All]**",
        "operationId": "epPostCreateExerciseProductionId",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "exerciseProduction",
            "in": "body",
            "description": "The exercise production's data",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["production_data"],
              "properties": {
                "ex_id": {
                  "type": "integer"
                },
                "ps_id": {
                  "type": "integer"
                },
                "user_id": {
                  "type": "integer"
                },
                "is_final": {
                  "type": "boolean"
                },
                "processing_log": {
                  "type": "string"
                },
                "working_time": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "exerciseProductionFiles",
            "in": "formData",
            "description": "The exercise production's files",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "201": {
            "description": "The new exercise production's id",
            "schema": {
              "type": "object",
              "required": ["ep_id"],
              "properties": {
                "ep_id": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Error: Missing parameters"
          },
          "401": {
            "description": "Error: Authorization required"
          },
          "500": {
            "description": "Error: Internal Server Error"
          }
        }
      }
    },
...
    "/api/student-statement": {
      "post": {
        "tags": ["StudentStatement"],
        "summary": "Add a user's statement to a PlageSession's exercise",
        "description": "Add a user's statement to a PlageSession's exercise **Authorized roles [Teacher, Admin]**",
        "operationId": "studentStatementPostExerciseUserId",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "StudentStatement",
            "in": "body",
            "description": "Add a PlageSession's exercise to a user's body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["ps_id", "user_id", "ex_id", "deadline_date", "is_sended"],
              "properties": {
                "ps_id": {
                  "type": "integer"
                },
                "user_id": {
                  "type": "integer"
                },
                "ex_id": {
                  "type": "integer"
                },
                "availability_date": {
                  "format": "date",
                  "type": "string"
                },
                "deadline_date": {
                  "format": "date",
                  "type": "string"
                },
                "is_sended": {
                  "type": "boolean"
                }
              }
            }
          },
          {
            "name": "file",
            "in": "formData",
            "description": "The associated file",
            "required": true,
            "type": "file"
          },
          {
            "name": "statement",
            "in": "formData",
            "description": "The exercise's statement",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "201": {
            "description": "Exercise added to user",
            "schema": {
              "type": "object",
              "properties": {
                "ps_id": {
                  "type": "integer"
                },
                "user_id": {
                  "type": "integer"
                },
                "ex_id": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Authorization required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/user": {
      "post": {
        "tags": ["User"],
        "summary": "Create a User",
        "description": "Create a User **Authorized roles [All]**",
        "operationId": "userPostUserId",
        "consumes": [],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "description": "Register User's body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "firstname",
                "lastname",
                "student_number",
                "email",
                "password",
                "organization",
                "country",
                "locale",
                "tdgroup"
              ],
              "properties": {
                "firstname": {
                  "type": "string"
                },
                "lastname": {
                  "type": "string"
                },
                "student_number": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "organization": {
                  "type": "string"
                },
                "country": {
                  "type": "string"
                },
                "locale": {
                  "format": "loc",
                  "type": "string"
                },
                "tdgroup": {
                  "type": "string"
                },
                "avatar": {
                  "format": "bytea",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Authorization required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/user/login": {
      "post": {
        "tags": ["User"],
        "summary": "Log in a user",
        "description": "Log in a user **Authorized roles [All]**",
        "operationId": "userLoginUserId",
        "consumes": [],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "description": "Log in User's body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "password"],
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User logged in",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "integer"
                },
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Authorization required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/logout": {
      "delete": {
        "tags": ["User"],
        "summary": "Log out the current user",
        "description": "Log out the current user **Authorized roles [All]**",
        "operationId": "userDeleteLogoutId",
        "consumes": [],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "User logged out"
          },
          "401": {
            "description": "Authorization required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
...
}